Rename router.py -> generator_router.py
Pure rename of the generator-router module and its test
(router.py -> generator_router.py, test_router.py ->
test_generator_router.py) plus the import-path updates in trainer.py,
config_registry.py, and the tests. No logic changes.
